xzdec: placate scan-build (WIP)
scan-build reports a possible dangling reference in `uncompress()`, where
`strm->avail_in` points to no-longer-valid (stack) `out_buf` on return.

This shouldn't be a problem in reality because we should be in LZMA_FINISH
at this point and therefore callers shouldn't do anything with `strm->avail_in`,
but it's simple enough for us to NULL-out `out_buf`, so let's do it.

Ditto for the case with LZMA_STREAM_END.

TODO: check this reasoning is right!